feat: rebrand cryptozombies to cryptopets
