+++ contracts/solana/cryptopets/.gitignore
