+++ contracts/solana/cryptopets/Cargo.toml
