+++ contracts/solana/cryptopets/package.json
