+++ contracts/solana/cryptopets/programs/cryptopets/Cargo.toml
@@ -1,12 +1,12 @@
 [package]
-name = "cryptozombies"
+name = "cryptopets"
 version = "0.1.0"
 description = "Created with Anchor"
 edition = "2021"
 
 [lib]
 crate-type = ["cdylib", "lib"]
-name = "cryptozombies"
+name = "cryptopets"
 
 [features]
 default = []
