+++ contracts/solana/cryptopets/programs/cryptopets/src/errors.rs
