+++ contracts/solana/cryptopets/programs/cryptopets/src/instructions/pause.rs
@@ -1,0 +1,7 @@
+use anchor_lang::prelude::*;
+
+pub fn handler(ctx: Context<crate::Pause>) -> Result<()> {
+    let global_state = &mut ctx.accounts.global_state;
+    global_state.paused = true;
+    Ok(())
+}
