+++ contracts/solana/cryptopets/programs/cryptopets/src/instructions/unpause.rs
@@ -1,0 +1,7 @@
+use anchor_lang::prelude::*;
+
+pub fn handler(ctx: Context<crate::Unpause>) -> Result<()> {
+    let global_state = &mut ctx.accounts.global_state;
+    global_state.paused = false;
+    Ok(())
+}
