+++ contracts/solana/cryptopets/scripts/inject-ngrok.ts
