+++ contracts/solana/cryptopets/tsconfig.json
