--- contracts/solana/cryptozombies/programs/cryptozombies/src/lib.rs
@@ -1,52 +1,0 @@
-pub mod errors;
-pub mod instructions;
-pub mod state;
-
-use anchor_lang::prelude::*;
-use instructions::{
-    create_starter_zombie,
-    initialize,
-    level_up,
-    rename_zombie,
-    pause,
-    unpause,
-};
-
-declare_id!("DVqmS2WV2epdEncuweNryuqyhbzLXtEQhnxDgQt7ccC6");
-
-#[program]
-pub mod cryptozombies {
-    use super::*;
-
-    pub fn initialize(
-        ctx: Context<Initialize>,
-        level_up_fee_lamports: u64,
-    ) -> Result<()> {
-        initialize::handler(ctx, level_up_fee_lamports)
-    }
-
-    pub fn create_starter_zombie(
-        ctx: Context<CreateStarterZombie>,
-        name: String,
-        dna: u64,
-        rarity: u8,
-    ) -> Result<()> {
-        create_starter_zombie::handler(ctx, name, dna, rarity)
-    }
-
-    pub fn level_up(ctx: Context<LevelUp>) -> Result<()> {
-        level_up::handler(ctx)
-    }
-
-    pub fn rename_zombie(ctx: Context<RenameZombie>, name: String) -> Result<()> {
-        rename_zombie::handler(ctx, name)
-    }
-
-    pub fn pause(ctx: Context<Pause>) -> Result<()> {
-        pause::handler(ctx)
-    }
-
-    pub fn unpause(ctx: Context<Unpause>) -> Result<()> {
-        unpause::handler(ctx)
-    }
-}
