+++ package.json
@@ -12,7 +12,7 @@
     "eth:deploy": "pnpm --prefix contracts/ethereum deploy:inject",
     "eth:vrf:watch": "pnpm --prefix contracts/ethereum vrf:watch",
     "sol:docker": "docker-compose -f contracts/solana/docker-compose.yml up -d",
-    "sol:inject-ngrok": "pnpm --prefix contracts/solana/cryptozombies inject-ngrok",
+    "sol:inject-ngrok": "pnpm --prefix contracts/solana/cryptopets inject-ngrok",
     "fe:eth:local": "concurrently --kill-others-on-fail --names \"HH,ETH,VRF,BE,FE\" \"pnpm eth:node\" \"pnpm eth:deploy\" \"pnpm eth:vrf:watch\" \"pnpm dev:be\" \"pnpm dev:fe\"",
     "mobile:eth:local": "concurrently --kill-others-on-fail --names \"HH,ETH,VRF,BE,MB\" \"pnpm eth:node\" \"pnpm eth:deploy\" \"pnpm eth:vrf:watch\" \"pnpm dev:be\" \"pnpm dev:mobile\"",
     "fe:sol:local": "concurrently --kill-others-on-fail --names \"BE,FE,SOL\" \"pnpm dev:be\" \"pnpm dev:fe\" \"pnpm sol:docker && pnpm sol:inject-ngrok\"",
